+++ enterprise/internal/codeintel/uploads/iface.go
@@ -6,16 +6,55 @@ import (
 	"time"
 
 	"github.com/grafana/regexp"
+	"github.com/sourcegraph/log"
 
+	codeinteltypes "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/shared/types"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/uploads/internal/background"
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/uploads/shared"
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/protocol"
+	"github.com/sourcegraph/sourcegraph/internal/observation"
+	"github.com/sourcegraph/sourcegraph/internal/uploadstore"
 )
 
-type UploadService = background.UploadService
+type UploadService interface {
+	// Commits
+	GetStaleSourcedCommits(ctx context.Context, minimumTimeSinceLastCheck time.Duration, limit int, now time.Time) (_ []shared.SourcedCommits, err error)
+	DeleteSourcedCommits(ctx context.Context, repositoryID int, commit string, maximumCommitLag time.Duration, now time.Time) (uploadsUpdated int, uploadsDeleted int, err error)
+	UpdateSourcedCommits(ctx context.Context, repositoryID int, commit string, now time.Time) (uploadsUpdated int, err error)
+	BackfillCommittedAtBatch(ctx context.Context, batchSize int) (err error)
+
+	// Uploads
+	GetUploads(ctx context.Context, opts shared.GetUploadsOptions) (uploads []codeinteltypes.Upload, totalCount int, err error)
+	SoftDeleteExpiredUploads(ctx context.Context, batchSize int) (int, error)
+	SoftDeleteExpiredUploadsViaTraversal(ctx context.Context, maxTraversal int) (int, error)
+	DeleteUploadsWithoutRepository(ctx context.Context, now time.Time) (_ map[int]int, err error)
+	DeleteUploadsStuckUploading(ctx context.Context, uploadedBefore time.Time) (_ int, err error)
+	DeleteLsifDataByUploadIds(ctx context.Context, bundleIDs ...int) (err error)
+	HardDeleteUploadsByIDs(ctx context.Context, ids ...int) error
+	HandleRawUpload(ctx context.Context, logger log.Logger, upload codeinteltypes.Upload, uploadStore uploadstore.Store, trace observation.TraceLogger) (requeued bool, err error)
+	HandleExpiredUploadsBatch(ctx context.Context, metrics *background.ExpirationMetrics, cfg background.ExpirerConfig) (err error)
+
+	// Commitgraph
+	UpdateAllDirtyCommitGraphs(ctx context.Context, maxAgeForNonStaleBranches time.Duration, maxAgeForNonStaleTags time.Duration) (err error)
+
+	// Repositories
+	GetDirtyRepositories(ctx context.Context) (_ map[int]int, err error)
+	GetRepositoriesMaxStaleAge(ctx context.Context) (_ time.Duration, err error)
+	SetRepositoriesForRetentionScan(ctx context.Context, processDelay time.Duration, limit int) (_ []int, err error)
+
+	// Audit logs
+	DeleteOldAuditLogs(ctx context.Context, maxAge time.Duration, now time.Time) (count int, err error)
+
+	// Utils
+	FrontendReconcileCandidates(ctx context.Context, batchSize int) ([]int, error)
+	CodeIntelDBReconcileCandidates(ctx context.Context, batchSize int) ([]int, error)
+	GetDumpsByIDs(ctx context.Context, ids []int) ([]codeinteltypes.Dump, error)
+	IDsWithMeta(ctx context.Context, ids []int) ([]int, error)
+}
 
 type Locker = background.Locker
